api/requirements.txt
Django==2.1.2
djangorestframework==3.8.2
django-rest-knox==3.1.5
pytz==2018.5
python-dateutil